Starting the CentOS install recipe.
The default installation documentation for GitLab CI only covers the
the Debian distribution. This documentation aims to add a recipe to get
GitLab CI running on the CentOS/RHEL 6 linux distribution.

The recipe is largely based off the Debian installation documention, so
this commit merely copies the Debian installation instructions to the
recipes directory so it can be pushed back to GitHub where the markdown
can be edited more easily.